backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-dotenv
pydantic
langchain
langchain-google-genai
langchain-pinecone
langchain-text-splitters